♻️ Move orchestration flows out of transform into dedicated module
auto_transform and smart_transform are higher-level flows that coordinate
detection, LLM generation, review, and registration. Splitting them into
orchestration.py clarifies that transform.py is purely the low-level
expression engine. Shared fake-LLM fixtures move to tests/conftest.py.
