+++ src/schemashift/transform.py
@@ -1,23 +1,14 @@
 """Core transform engine: apply a FormatConfig to a file."""
 
-from collections.abc import Callable
 from pathlib import Path
-from typing import TYPE_CHECKING
 
 import polars as pl
 
 from schemashift.dsl import parse_and_compile
 from schemashift.dtypes import DTYPE_MAP
-from schemashift.errors import DSLRuntimeError, DSLSyntaxError, FormatDetectionError, ReviewRejectedError
-from schemashift.models import ColumnMapping, FormatConfig, ReaderConfig
+from schemashift.errors import DSLSyntaxError
+from schemashift.models import ColumnMapping, FormatConfig
 from schemashift.readers import read_file
-from schemashift.registry import Registry
-
-if TYPE_CHECKING:
-    from langchain_core.language_models import BaseChatModel
-
-    from schemashift.target_schema import TargetSchema
-
 
 # ---------------------------------------------------------------------------
 # Public API
@@ -27,7 +18,6 @@ if TYPE_CHECKING:
 def transform(
     path: str | Path,
     config: FormatConfig,
-    reader_config: ReaderConfig | None = None,
 ) -> pl.LazyFrame:
     """Apply a FormatConfig to a file and return a transformed LazyFrame.
 
@@ -41,8 +31,6 @@ def transform(
     Args:
         path: Path to the source file.
         config: FormatConfig describing how to map columns.
-        reader_config: Optional low-level reader options (overrides
-            ``config.reader`` when provided).
 
     Returns:
         A :class:`polars.LazyFrame` with the transformed columns.
@@ -51,8 +39,7 @@ def transform(
         DSLRuntimeError: When a DSL expression fails to evaluate.
         ReaderError: When the file cannot be read.
     """
-    effective_reader = reader_config if reader_config is not None else config.reader
-    lf = read_file(path, effective_reader)
+    lf = read_file(path, config.reader)
     expressions = _build_expressions(config)
 
     if config.drop_unmapped:
@@ -107,131 +94,6 @@ def dry_run(config: FormatConfig, path: str | Path, n_rows: int = 10) -> pl.Data
     return result
 
 
-def auto_transform(
-    path: str | Path,
-    registry: Registry,
-    reader_config: ReaderConfig | None = None,
-) -> pl.LazyFrame:
-    """Auto-detect the format from the registry and transform the file.
-
-    Args:
-        path: Path to the source file.
-        registry: Registry to search for a matching config.
-        reader_config: Optional reader configuration forwarded to the file reader.
-
-    Returns:
-        A transformed :class:`polars.LazyFrame`.
-
-    Raises:
-        FormatDetectionError: When no config matches the file's columns.
-        AmbiguousFormatError: When multiple configs match.
-    """
-    from schemashift.detection import detect_format
-    from schemashift.readers import read_header
-
-    columns = read_header(path, reader_config)
-    config = detect_format(columns, registry)
-
-    if config is None:
-        raise FormatDetectionError(
-            f"No registered config matches the columns found in '{path}'. Columns present: {columns}"
-        )
-
-    return transform(path, config, reader_config)
-
-
-def smart_transform(
-    path: str | Path,
-    registry: Registry,
-    target_schema: "TargetSchema | None" = None,
-    llm: "BaseChatModel | None" = None,
-    review_fn: Callable[[FormatConfig, pl.DataFrame], FormatConfig | None] | None = None,
-    auto_register: bool = False,
-    example_configs: list[FormatConfig] | None = None,
-    max_retries: int = 2,
-    n_sample_rows: int = 15,
-    user_prompt: str | None = None,
-    reader_config: ReaderConfig | None = None,
-) -> pl.LazyFrame:
-    """Full detect-or-generate flow.
-
-    1. Try auto-detect from registry.
-    2. If miss and LLM available: generate config.
-    3. If review_fn provided: pass config + sample to reviewer.
-    4. If auto_register: save to registry.
-    5. Apply config; optionally validate against target_schema.
-
-    Args:
-        path: Source file path.
-        registry: Registry to search and optionally register to.
-        target_schema: Required for LLM generation and output validation.
-        llm: LangChain BaseChatModel.
-        review_fn: callback(config, sample_df) -> config | None. None = reject.
-        auto_register: Register LLM-generated config automatically.
-        example_configs: Example configs for LLM prompt.
-        max_retries: Max LLM retries.
-        n_sample_rows: Rows to sample for LLM.
-        user_prompt: Additional context for the LLM (e.g. unit conventions).
-        reader_config: Optional reader configuration forwarded to all file reads.
-
-    Returns:
-        Transformed pl.LazyFrame.
-
-    Raises:
-        FormatDetectionError: No match and no LLM, or review_fn rejected.
-        ValueError: LLM needed but target_schema not provided.
-        LLMGenerationError: LLM fails after all retries.
-    """
-    from schemashift.detection import detect_format
-    from schemashift.readers import read_header
-
-    # Try registry
-    columns = read_header(path, reader_config)
-    config = detect_format(columns, registry)
-
-    if config is not None:
-        lf = transform(path, config, reader_config)
-        if target_schema is not None:
-            target_schema.validate_lazy(lf)
-        return lf
-
-    # No match — need LLM
-    if llm is None:
-        raise FormatDetectionError(
-            f"No registered config matches '{path}' and no LLM is configured. File columns: {columns}"
-        )
-    if target_schema is None:
-        raise ValueError("target_schema is required for LLM config generation")
-
-    from schemashift.llm import generate_config as _llm_generate
-
-    config = _llm_generate(
-        path=str(path),
-        target_schema=target_schema,
-        llm=llm,
-        example_configs=example_configs,
-        max_retries=max_retries,
-        n_sample_rows=n_sample_rows,
-        user_prompt=user_prompt,
-        reader_config=reader_config,
-    )
-
-    # Review callback
-    if review_fn is not None:
-        sample_df = dry_run(config, path, n_rows=10)
-        reviewed = review_fn(config, sample_df)
-        if reviewed is None:
-            raise ReviewRejectedError("Config was rejected by review_fn")
-        config = reviewed
-
-    if auto_register:
-        registry.register(config)
-
-    lf = transform(path, config, reader_config)
-    if target_schema is not None:
-        target_schema.validate_lazy(lf)
-    return lf
-
 
 # ---------------------------------------------------------------------------
 # Private helpers
@@ -254,14 +116,7 @@ def _mapping_to_expr(mapping: ColumnMapping) -> pl.Expr:
     if mapping.source is not None:
         expr = pl.col(mapping.source).alias(mapping.target)
     elif mapping.expr is not None:
-        try:
-            compiled = parse_and_compile(mapping.expr)
-        except DSLSyntaxError as exc:
-            raise DSLRuntimeError(
-                f"DSL syntax error for target '{mapping.target}': {exc}",
-                expression=mapping.expr,
-                target=mapping.target,
-            ) from exc
+        compiled = parse_and_compile(mapping.expr)
         expr = compiled.alias(mapping.target)
     elif mapping.has_constant():
         # constant (may be None, broadcasting nulls)
